+++ frontend/src/pages/xray/outbounds/transport/index.ts
@@ -4,3 +4,6 @@ export { default as WsForm } from './ws';
 export { default as GrpcForm } from './grpc';
 export { default as HttpUpgradeForm } from './httpupgrade';
 export { default as XhttpForm } from './xhttp';
+export { default as HysteriaForm } from './hysteria';
+export { default as SockoptForm } from './sockopt';
+export { default as MuxForm } from './mux';
